+++ packages/behandling-anke/package.json
@@ -9,7 +9,7 @@
     "@fpsak-frontend/kodeverk": "1.0.0",
     "@fpsak-frontend/konstanter": "1.0.0",
     "@fpsak-frontend/prosess-anke": "1.0.0",
-    "@fpsak-frontend/prosess-anke-merknader": "1.0.0",
+    "@fpsak-frontend/prosess-anke-trygderettsbehandling": "1.0.0",
     "@fpsak-frontend/prosess-anke-resultat": "1.0.0",
     "@fpsak-frontend/rest-api": "1.0.0",
     "@fpsak-frontend/rest-api-hooks": "1.0.0",
