--- packages/prosess-anke-merknader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './src/AnkeMerknaderProsessIndex';
