+++ packages/prosess-anke-resultat/src/components/BehandleResultatForm.tsx
@@ -133,7 +133,6 @@ const AnkeResultat: FunctionComponent<OwnPropsResultat & { alleKodeverk: {[key: 
 };
 
 interface OwnProps {
-  saveAnke: (data: any) => Promise<any>;
   previewCallback: (data: any) => Promise<any>;
   aksjonspunktCode: string;
   ankeVurderingVerdi?: Kodeverk;
