+++ packages/prosess-anke-trygderettsbehandling/index.ts
@@ -1,0 +1,1 @@
+export { default } from './src/AnkeTrygderettsbehandlingProsessIndex';
