+++ packages/prosess-anke-trygderettsbehandling/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@fpsak-frontend/prosess-anke-merknader",
+  "name": "@fpsak-frontend/prosess-anke-trygderettsbehandling",
   "version": "1.0.0",
   "module": "index.ts",
   "license": "MIT",
