+++ packages/prosess-anke-trygderettsbehandling/src/AnkeTrygderettsbehandlingProsessIndex.tsx
@@ -5,7 +5,8 @@ import kodeverkTyper from '@fpsak-frontend/kodeverk/src/kodeverkTyper';
 import { AnkeVurdering } from '@fpsak-frontend/types';
 import { StandardProsessFormProps } from '@fpsak-frontend/prosess-felles';
 
-import BehandleMerknaderForm from './components/BehandleMerknaderForm';
+import TrygderettsbehandlingForm from './components/TrygderettsbehandlingForm';
+import { BrevData } from './components/PreviewAnkeLink';
 import messages from '../i18n/nb_NO.json';
 
 const cache = createIntlCache();
@@ -17,23 +18,21 @@ const intl = createIntl({
 
 interface OwnProps {
   ankeVurdering: AnkeVurdering;
-  saveAnke: (data: any) => Promise<any>;
-  previewCallback: (data: any) => Promise<any>;
+  previewCallback: (data: BrevData) => Promise<any>;
 }
 
-const AnkeMerknaderProsessIndex: FunctionComponent<OwnProps & StandardProsessFormProps> = ({
+const AnkeTrygderettsbehandlingProsessIndex: FunctionComponent<OwnProps & StandardProsessFormProps> = ({
   behandling,
   ankeVurdering,
   aksjonspunkter,
   submitCallback,
   isReadOnly,
   readOnlySubmitButton,
-  saveAnke,
   previewCallback,
   alleKodeverk,
 }) => (
   <RawIntlProvider value={intl}>
-    <BehandleMerknaderForm
+    <TrygderettsbehandlingForm
       behandlingId={behandling.id}
       behandlingVersjon={behandling.versjon}
       ankeVurderingResultat={ankeVurdering.ankeVurderingResultat}
@@ -42,11 +41,10 @@ const AnkeMerknaderProsessIndex: FunctionComponent<OwnProps & StandardProsessFor
       sprakkode={behandling.sprakkode}
       readOnly={isReadOnly}
       readOnlySubmitButton={readOnlySubmitButton}
-      saveAnke={saveAnke}
       previewCallback={previewCallback}
       ankeOmgorArsaker={alleKodeverk[kodeverkTyper.ANKE_OMGJOER_AARSAK]}
     />
   </RawIntlProvider>
 );
 
-export default AnkeMerknaderProsessIndex;
+export default AnkeTrygderettsbehandlingProsessIndex;
