+++ packages/prosess-anke-trygderettsbehandling/src/components/FritekstAnkeMerknaderTextField.spec.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import { expect } from 'chai';
+
+import { intlMock } from '@fpsak-frontend/utils-test/src/intl-enzyme-test-helper';
+import { TextAreaField } from '@fpsak-frontend/form';
+
+import shallowWithIntl from '../../i18n/intl-enzyme-test-helper-prosess-anke-trygderettsbehandling';
+import FritekstAnkeMerknaderTextField from './FritekstAnkeMerknaderTextField';
+
+describe('<TrygderettsbehandlingForm>', () => {
+  const sprakkode = {
+    kode: 'NB',
+    kodeverk: '',
+  };
+
+  it('skal rendre komponent korrekt', () => {
+    const wrapper = shallowWithIntl(
+      <FritekstAnkeMerknaderTextField.WrappedComponent
+        intl={intlMock}
+        sprakkode={sprakkode}
+        readOnly={false}
+      />,
+    );
+
+    expect(wrapper.find(TextAreaField)).has.length(1);
+  });
+});
