+++ packages/prosess-anke-trygderettsbehandling/src/components/PreviewAnkeLink.tsx
@@ -8,7 +8,7 @@ import { Kodeverk } from '@fpsak-frontend/types';
 
 import styles from './previewAnkeLink.less';
 
-const getBrevKode = (ankeVurdering?: Kodeverk) => {
+const getBrevKode = (ankeVurdering?: Kodeverk): string => {
   switch (ankeVurdering.kode) {
     case ankeVurderingType.ANKE_OMGJOER:
       return dokumentMalType.ANKE_VEDTAK_OMGJORING;
@@ -19,7 +19,13 @@ const getBrevKode = (ankeVurdering?: Kodeverk) => {
   }
 };
 
-const getBrevData = (ankeVurdering?: Kodeverk, fritekstTilBrev?: string) => {
+export type BrevData = {
+  fritekst: string;
+  mottaker: string;
+  dokumentMal?: string;
+}
+
+const getBrevData = (ankeVurdering?: Kodeverk, fritekstTilBrev?: string): BrevData => {
   const data = {
     fritekst: fritekstTilBrev || '',
     mottaker: '',
@@ -29,7 +35,7 @@ const getBrevData = (ankeVurdering?: Kodeverk, fritekstTilBrev?: string) => {
 };
 
 interface OwnProps {
-  previewCallback: (data: any) => Promise<any>;
+  previewCallback: (data: BrevData) => Promise<any>;
   aksjonspunktCode: string;
   fritekstTilBrev?: string;
   ankeVurdering?: Kodeverk;
@@ -42,10 +48,6 @@ const PreviewAnkeLink: FunctionComponent<OwnProps> = ({
   ankeVurdering,
   readOnly = false,
 }) => {
-  const previewMessage = (e: any) => {
-    e.preventDefault();
-    previewCallback(getBrevData(ankeVurdering, fritekstTilBrev));
-  };
   if (readOnly) {
     return (
       <span
@@ -55,6 +57,10 @@ const PreviewAnkeLink: FunctionComponent<OwnProps> = ({
       </span>
     );
   }
+  const previewMessage = (e: React.MouseEvent | React.KeyboardEvent) => {
+    e.preventDefault();
+    previewCallback(getBrevData(ankeVurdering, fritekstTilBrev));
+  };
   return (
     <a
       href=""
